PPM-8636: refactor
