+++ cmd/postgres_exporter/postgres_exporter_PMM-8646_test.go
@@ -1,50 +1,37 @@
 package main
 
 import (
-	"bufio"
-	"errors"
 	"fmt"
-	"github.com/percona/exporter_shared"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/common/version"
-	"go.uber.org/atomic"
-	"net"
-	"net/http"
 	"regexp"
-	"strings"
 	"testing"
-	"time"
 )
 
 var (
-	exporterPort, _         = GetFreePort()
 	//exporterPort         = 10000
-	exporterMetricsEndpoint = fmt.Sprintf("http://localhost:%d/metrics", exporterPort)
-	pgPort                  = 55333 // you should not have PG be running on this port
-	dataSources             = []string{
+	pgPort      = 55333 // you should not have PG be running on this port
+	dataSources = []string{
 		fmt.Sprintf("postgresql://root:root@localhost:%d/postgres", pgPort),
 	}
 )
 
 func TestRespectConnectivityErrorsOnMetricsQuerying(t *testing.T) {
-	pingWorks := atomic.NewBool(false)
-
-	exporterReady := make(chan interface{})
-	go runExporter(
-		exporterReady,
+	it := NewPGIntegrationTest()
+	it.RunExporter(
 		CustomServerPing(func(s *Server) error {
-			if pingWorks.Load() {
+			if it.pingWorks.Load() {
 				return nil
 			} else {
 				return &ErrorConnectToServer{Msg: "connection failure"}
 			}
 		}),
-		CustomGetServerRetry(&CustomGetServerRetryFactory{}),
-	)
-	<-exporterReady
+		CustomGetServerRetry(&CustomGetServerRetryFactory{}))
+	go func() {
+		it.StopExporter()
+	}()
+	it.SetPingWorks(false)
 
 	fetchMetrics := func() map[string]string {
-		result, err := _fetchMetrics(
+		result, err := it.FetchMetrics(
 			"pg_up",
 			regexp.MustCompile(fmt.Sprintf("pg_stat_database_(?P<key>tup_fetched).*postgres.*localhost:%d\"} (?P<value>.*)", pgPort)),
 		)
@@ -64,7 +51,7 @@ func TestRespectConnectivityErrorsOnMetricsQuerying(t *testing.T) {
 	}
 
 	// simulating case when "db ping" is successful and right after that db is not available
-	pingWorks.Store(true)
+	it.SetPingWorks(true)
 
 	metrics = fetchMetrics()
 	if metrics["pg_up"] != "0" {
@@ -73,117 +60,4 @@ func TestRespectConnectivityErrorsOnMetricsQuerying(t *testing.T) {
 	if _, ok := metrics["tup_fetched"]; ok {
 		t.Fatalf("pg_up was reported as 0, but actual metric is present")
 	}
-}
-
-func runExporter(ready chan interface{}, opts ...ExporterOpt) {
-
-	exporter := NewExporter(dataSources, opts...)
-
-	defer func() {
-		exporter.servers.Close()
-	}()
-	prometheus.MustRegister(exporter)
-
-	psCollector := prometheus.NewProcessCollector(prometheus.ProcessCollectorOpts{})
-	goCollector := prometheus.NewGoCollector()
-
-	version.Branch = Branch
-	version.BuildDate = BuildDate
-	version.Revision = Revision
-	version.Version = VersionShort
-	prometheus.MustRegister(version.NewCollector("postgres_exporter"))
-
-	go exporter_shared.RunServer("PostgreSQL", fmt.Sprintf(":%d", exporterPort), "/metrics", newHandler(map[string]prometheus.Collector{
-		"exporter":         exporter,
-		"standard.process": psCollector,
-		"standard.go":      goCollector,
-	}))
-	for {
-		_, err := http.Get(exporterMetricsEndpoint)
-		if err == nil {
-			ready <- nil
-			break
-		}
-		time.Sleep(100 * time.Microsecond)
-	}
-}
-
-func _fetchMetrics(metrics ...interface{}) (map[string]string, error) {
-	result := make(map[string]string)
-
-	resp, err := http.Get(exporterMetricsEndpoint)
-	if err != nil {
-		return nil, err
-	}
-
-	sc := bufio.NewScanner(resp.Body)
-	for sc.Scan() {
-		next := sc.Text()
-		for _, metric := range metrics {
-			metricStr, ok := metric.(string)
-			if ok {
-				starts := strings.HasPrefix(next, metricStr)
-				if starts {
-					keyLen := len(metricStr) + 1
-					value := next[keyLen:]
-					result[metricStr] = value
-				}
-
-				continue
-			}
-
-			metricsRE, ok := metric.(*regexp.Regexp)
-			yes := strings.HasPrefix(next, "pg_stat_database")
-			if ok && yes {
-				matches := metricsRE.FindAllStringSubmatch(next, -1)
-				if matches != nil {
-					key := matches[0][metricsRE.SubexpIndex("key")]
-					value := matches[0][metricsRE.SubexpIndex("value")]
-					result[key] = value
-				}
-			}
-
-		}
-	}
-
-	return result, nil
-}
-
-// --- utils
-
-type CustomGetServerRetryFactory struct {
-}
-
-func (c *CustomGetServerRetryFactory) Create() GetServerRetry {
-	return &RetryStrategy{
-	}
-}
-
-type RetryStrategy struct {
-	failed bool
-}
-
-func (r *RetryStrategy) CanRetry() (yes bool, lastError error) {
-	if r.failed {
-		return false, errors.New("<test error>")
-	}
-	return true, nil
-}
-
-func (r *RetryStrategy) Fail(e error) {
-	r.failed = true
-}
-
-func GetFreePort() (int, error) {
-	addr, err := net.ResolveTCPAddr("tcp", "localhost:0")
-	if err != nil {
-		return 0, err
-	}
-
-	l, err := net.ListenTCP("tcp", addr)
-	if err != nil {
-		return 0, err
-	}
-	defer l.Close()
-	return l.Addr().(*net.TCPAddr).Port, nil
 }
