+++ docker-compose.yml
@@ -28,6 +28,5 @@ services:
     environment:
       - POSTGRES_USER=root
       - POSTGRES_PASSWORD=root
-    volumes:
-      - ./testdata/ssl/server:/ssl
+#    volumes:
 #      - ./pg_data:/var/lib/postgresql/data/
