+++ testdata/Dockerfile
@@ -3,6 +3,6 @@ ARG POSTGRESQL_IMAGE
 FROM ${POSTGRESQL_IMAGE}
 
 
-COPY testdata/ssl /ssl
-RUN chown 999:999 /ssl/* \
-    && chmod 0600 /ssl/*
+COPY testdata/ssl/server/* /ssl/
+RUN chown 999:999 /ssl/* && \
+    chmod 0600 /ssl/*
